refactor folders to seperate tests
