+++ .github/workflows/ci.yml
@@ -30,4 +30,4 @@ jobs:
     - name: Test with pytest
       run: |
         pip install pytest
-        pytest mitiq/tests/test_zne_qiskit.py
+        pytest mitiq/qiskit/tests
